Refactor ChannelI18nUtil to expose service directly
